backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.10.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
deepseek==1.0.0
faiss-cpu==1.10.0
filelock==3.18.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.56.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httpcore==1.0.7
httpx==0.25.2
huggingface-hub==0.29.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
langchain==0.0.1
langdetect==1.0.9
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
ollama==0.1.0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
priority==1.3.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.8.2
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
requests==2.26.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.5.0
sentence-transformers==2.1.0
sentencepiece==0.2.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.0
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
w3lib==2.3.1
Werkzeug==2.0.3
wordcloud==1.9.4
zope.interface==7.2
